feat: added price to performed tasks, added price calculations and number formatters, upgraded readonlyfield to allow more precise control

+++ backend/src/main/java/cz/zcu/kiv/caretracker/entity/PerformedTask.java
@@ -45,7 +45,7 @@ public class PerformedTask {
     private LocalDateTime date;
 
     @Column(name="unit_count", nullable = false)
-    private Integer unitCount;
+    private Double unitCount;
 
     @Column(columnDefinition = "TEXT")
     private String notes;
